improve GSO tests
