+++ sys_conn_helper_nonlinux_test.go
@@ -1,0 +1,7 @@
+//go:build !linux
+
+package quic
+
+import "errors"
+
+var errGSO = errors.New("fake GSO error")
